PROJECT/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.1
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
cryptography==41.0.0
pytest==7.4.0
python-dateutil==2.8.2
haversine==2.7.0